test: shard tui gateway s1 clusters
