+++ tests/test_tui_gateway_server.py
@@ -38,127 +38,6 @@ def _neuter_agent_prewarm_timer(request, monkeypatch):
     yield
 
 
-def test_session_slot_is_claimed_on_first_turn_not_on_create(monkeypatch, tmp_path):
-    home = tmp_path / ".hermes"
-    home.mkdir()
-    (home / "config.yaml").write_text("max_concurrent_sessions: 1\n", encoding="utf-8")
-    token = set_hermes_home_override(home)
-
-    def _clear_server_sessions():
-        for session in list(server._sessions.values()):
-            server._teardown_session(session)
-        server._sessions.clear()
-
-    try:
-        server._cfg_cache = None
-        server._cfg_mtime = None
-        server._cfg_path = None
-        _clear_server_sessions()
-        monkeypatch.setattr(server, "_start_agent_build", lambda *args, **kwargs: None)
-        monkeypatch.setattr(server, "_completion_cwd", lambda params=None: str(tmp_path))
-
-        # Opening a chat must NOT take a slot. Every tile paint and every
-        # background reconnect-resume calls session.create, and an unprompted
-        # draft has no DB row and is filtered out of the sidebar — so a slot
-        # held here is invisible to the user while still starving the other
-        # surfaces that share this cap.
-        first = server._methods["session.create"]("r1", {"cols": 80})
-        second = server._methods["session.create"]("r2", {"cols": 80})
-        assert "result" in first and "result" in second
-        sid = first["result"]["session_id"]
-        other = second["result"]["session_id"]
-        assert active_session_registry_snapshot() == []
-
-        # The first turn is what claims the slot, and is re-entrant.
-        assert server._ensure_active_session_slot(sid, server._sessions[sid]) is None
-        assert server._ensure_active_session_slot(sid, server._sessions[sid]) is None
-        assert len(active_session_registry_snapshot()) == 1
-
-        blocked = server._ensure_active_session_slot(other, server._sessions[other])
-        assert "active session limit (1/1)" in blocked
-
-        closed = server._methods["session.close"]("r3", {"session_id": sid})
-        assert closed["result"]["closed"] is True
-        assert active_session_registry_snapshot() == []
-
-        assert server._ensure_active_session_slot(other, server._sessions[other]) is None
-    finally:
-        _clear_server_sessions()
-        server._cfg_cache = None
-        server._cfg_mtime = None
-        server._cfg_path = None
-        reset_hermes_home_override(token)
-
-
-def test_session_context_uses_session_cwd(monkeypatch, tmp_path):
-    """Desktop/TUI sessions must pin the agent cwd per session.
-
-    The gateway process itself is often launched from apps/desktop in dev, so
-    falling back to os.getcwd() makes agents answer from the desktop app folder
-    even when the sidebar/session cwd is a real project.
-    """
-    from agent.runtime_cwd import resolve_agent_cwd
-
-    sid = "cwd-sid"
-    session_key = "cwd-key"
-    project = tmp_path / "project"
-    project.mkdir()
-    (project / ".git").mkdir()
-    launcher = tmp_path / "apps" / "desktop"
-    launcher.mkdir(parents=True)
-
-    server._sessions[sid] = {"session_key": session_key, "cwd": str(project)}
-    monkeypatch.delenv("TERMINAL_CWD", raising=False)
-    monkeypatch.chdir(launcher)
-
-    tokens = server._set_session_context(session_key)
-    try:
-        assert resolve_agent_cwd() == project
-    finally:
-        server._clear_session_context(tokens)
-        server._sessions.pop(sid, None)
-
-
-def test_handoff_fail_marks_only_inflight_rows(monkeypatch):
-    class DbContext:
-        def __init__(self, db):
-            self.db = db
-
-        def __enter__(self):
-            return self.db
-
-        def __exit__(self, *_args):
-            return False
-
-    class FakeDb:
-        def __init__(self, state):
-            self.state = state
-            self.failed_with = None
-
-        def get_handoff_state(self, _key):
-            return {"state": self.state, "platform": "telegram", "error": None}
-
-        def fail_handoff(self, _key, error):
-            self.failed_with = error
-            self.state = "failed"
-
-    sid = "rt-handoff"
-    server._sessions[sid] = {"session_key": "stored-handoff"}
-    try:
-        pending = FakeDb("pending")
-        monkeypatch.setattr(server, "_session_db", lambda _session: DbContext(pending))
-        result = server._methods["handoff.fail"]("r1", {"session_id": sid, "error": "timed out"})
-        assert result["result"] == {"failed": True, "state": "failed"}
-        assert pending.failed_with == "timed out"
-
-        completed = FakeDb("completed")
-        monkeypatch.setattr(server, "_session_db", lambda _session: DbContext(completed))
-        result = server._methods["handoff.fail"]("r2", {"session_id": sid, "error": "late timeout"})
-        assert result["result"] == {"failed": False, "state": "completed"}
-        assert completed.failed_with is None
-    finally:
-        server._sessions.pop(sid, None)
-
 
 def test_dashboard_process_isolation_config_defaults_without_default_merge(monkeypatch):
     """tui_gateway.server::_load_cfg is raw YAML, so defaults live at read site."""
@@ -560,27 +439,6 @@ def test_prompt_submit_golden_transcript_matches_flag_off_and_on(monkeypatch):
             server._sessions.pop("sid", None)
 
     assert run_flag_on() == run_flag_off()
-
-
-def test_session_context_explicit_cwd_for_ephemeral_task(monkeypatch, tmp_path):
-    """Background/preview tasks use ephemeral ids absent from `_sessions`, so the
-    parent workspace is passed explicitly; it must pin instead of clearing back
-    to the gateway launch dir."""
-    from agent.runtime_cwd import resolve_agent_cwd
-
-    project = tmp_path / "project"
-    project.mkdir()
-    launcher = tmp_path / "apps" / "desktop"
-    launcher.mkdir(parents=True)
-
-    monkeypatch.delenv("TERMINAL_CWD", raising=False)
-    monkeypatch.chdir(launcher)
-
-    tokens = server._set_session_context("bg_deadbe", cwd=str(project))
-    try:
-        assert resolve_agent_cwd() == project
-    finally:
-        server._clear_session_context(tokens)
 
 
 def _write_profile_cfg(home: Path, cwd: str | None) -> Path:
